refactor(role): rename files
