+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UserModule } from './user/user.module';
 import { GroupModule } from './group/group.module';
-import { RolesModule } from './roles/roles.module';
+import { RolesModule } from './role/role.module';
 import { PaymentModule } from './payment/payment.module';
 import { PassModule } from './pass/pass.module';
 import { PhoneModule } from './phone/phone.module';
