+++ src/role/entities/role.entity.ts
