+++ src/role/role.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { RolesController } from './roles.controller';
-import { RolesService } from './roles.service';
+import { RolesController } from './role.controller';
+import { RolesService } from './role.service';
 
 describe('RolesController', () => {
   let controller: RolesController;
