+++ src/role/role.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { RolesService } from './roles.service';
+import { RolesService } from './role.service';
 
 describe('RolesService', () => {
   let service: RolesService;
